Only allow users to delete tasks they own.

+++ task-list/app/Http/Controllers/TaskController.php
@@ -58,9 +58,15 @@ class TaskController extends Controller
     }
 
     /**
-     * Delete Task
+     * Destroy the given task.
+     * 
+     * @param   Request $request
+     * @param   Task    $task
+     * @return  Response
      */
-    function destroy(Task $task) {
+    function destroy(Request $request, Task $task) {
+        $this->authorize('destroy', $task);
+        
         $task->delete();
 
         return redirect('/tasks');
